+++ Gobchat.App/Core/Config/ProfilePropertyChangedEventArgs.cs
@@ -1,0 +1,31 @@
+﻿/*******************************************************************************
+ * Copyright (C) 2019 MarbleBag
+ *
+ * This program is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU Affero General Public License as published by the Free
+ * Software Foundation, version 3.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ *******************************************************************************/
+
+using System;
+
+namespace Gobchat.Core.Config
+{
+    public sealed class ProfilePropertyChangedEventArgs : EventArgs
+    {
+        public string PropertyKey { get; }
+        public string ProfileId { get; }
+        public bool IsActiveProfile { get; }
+
+        public ProfilePropertyChangedEventArgs(string propertyKey, string profileId, bool isActiveProfile)
+        {
+            PropertyKey = propertyKey;
+            ProfileId = profileId;
+            IsActiveProfile = isActiveProfile;
+        }
+    }
+}